Fix a bug that was causing searches to not fire when simple search run from the search page.

+++ src/main/java/mil/dds/anet/resources/PositionResource.java
@@ -195,6 +195,7 @@ public class PositionResource implements IGraphQLResource {
 	@RolesAllowed("SUPER_USER")
 	public Response deletePersonFromPosition(@Auth Person user, @PathParam("id") int positionId) {
 		Position pos = dao.getById(positionId);
+		if (pos == null) { return Response.status(Status.NOT_FOUND).build(); } 
 		AuthUtils.assertSuperUserForOrg(user, pos.getOrganization());
 
 		dao.removePersonFromPosition(pos);
